# swagger-output.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "User & Business API",
    "description": "API docs for user registration, OTP verification, and business creation.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "language": {
            "type": "string"
          },
          "mobileNumber": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "language",
          "mobileNumber",
          "password"
        ]
      },
...
    }
  }
}